requirements.txt
discord.py
asyncio
splite3